# frontend/src/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Fee Management API",
    "description": "RESTful API for school fee management system",
    "contact": {
      "name": "School Fee App Team",
      "email": "support@schoolfeeapp.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    },
    {
      "url": "http://localhost:8081",
      "description": "Docker Development"
    }
  ],
  "paths": {
    "/api/v1/students/{studentId}": {
      "get": {
        "tags": [
          "student-controller"
        ],
        "operationId": "getStudentDetails",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentDetailResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "student-controller"
        ],
        "operationId": "updateStudent",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUpdateStudentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fees/structures": {
      "get": {
        "tags": [
          "fee-controller"
        ],
        "operationId": "getFeeStructures",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ACTIVE"
            }
          },
          {
            "name": "termId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListFeeStructureResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fee-controller"
        ],
        "operationId": "createFeeStructure",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeeStructureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCreateFeeStructureResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/structures/{structureId}/assign": {
      "post": {
        "tags": [
          "fee-controller"
        ],
        "operationId": "assignFees",
        "parameters": [
          {
            "name": "structureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFeeAssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/staff": {
      "post": {
        "tags": [
          "user-management-controller"
        ],
        "operationId": "createStaff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCreateStaffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/parents": {
      "post": {
        "tags": [
          "user-management-controller"
        ],
        "operationId": "createParent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCreateParentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fees/students/{studentId}": {
      "get": {
        "tags": [
          "fee-controller"
        ],
        "operationId": "getStudentFees",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListStudentFeeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/dashboard": {
      "get": {
        "tags": [
          "fee-controller"
        ],
        "operationId": "getDashboard",
        "parameters": [
          {
            "name": "termId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "current"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseFeeDashboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "tags": [
          "user-management-controller"
        ],
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "userType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ACTIVE"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "minimum": 1
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "userId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseUserSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateStudentRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "currentClassId": {
            "type": "string",
            "format": "uuid"
          },
          "enrollmentStatus": {
            "type": "string"
          },
          "medicalNotes": {
            "type": "string"
          }
        }
      },
      "ApiResponseUpdateStudentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateStudentResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateStudentResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "admissionNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "currentClassId": {
            "type": "string",
            "format": "uuid"
          },
          "className": {
            "type": "string"
          },
          "enrollmentStatus": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateFeeStructureRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "applicableToClassIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeeItemRequest"
            }
          },
          "lateFeeConfig": {
            "$ref": "#/components/schemas/LateFeeConfig"
          }
        },
        "required": [
          "applicableToClassIds",
          "dueDate",
          "items",
          "name",
          "sessionId",
          "termId"
        ]
      },
      "FeeItemRequest": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "isMandatory": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "amount",
          "description"
        ]
      },
      "LateFeeConfig": {
        "type": "object",
        "properties": {
          "applyAfterDays": {
            "type": "integer",
            "format": "int32"
          },
          "percentageAmount": {
            "type": "number",
            "format": "double"
          },
          "flatAmount": {
            "type": "number"
          }
        }
      },
      "ApiResponseCreateFeeStructureResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CreateFeeStructureResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateFeeStructureResponse": {
        "type": "object",
        "properties": {
          "structureId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number"
          },
          "mandatoryAmount": {
            "type": "number"
          },
          "applicableClassCount": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedStudentCount": {
            "type": "integer",
            "format": "int32"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseFeeAssignmentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FeeAssignmentResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FeeAssignmentResponse": {
        "type": "object",
        "properties": {
          "structureId": {
            "type": "string",
            "format": "uuid"
          },
          "studentsAssigned": {
            "type": "integer",
            "format": "int32"
          },
          "totalExpectedAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "nextReminderDate": {
            "type": "string",
            "format": "date"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CreateStaffRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "phoneNumber",
          "roles",
          "userType"
        ]
      },
      "ApiResponseCreateStaffResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CreateStaffResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateStaffResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "schoolName": {
            "type": "string"
          },
          "temporaryPassword": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ChildLink": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "relationship": {
            "type": "string"
          },
          "isPrimaryContact": {
            "type": "boolean"
          }
        },
        "required": [
          "relationship",
          "studentId"
        ]
      },
      "CreateParentRequest": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,15}$"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChildLink"
            }
          }
        },
        "required": [
          "children",
          "firstName",
          "lastName",
          "phoneNumber"
        ]
      },
      "ApiResponseCreateParentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CreateParentResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseStudentDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/StudentDetailResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StudentDetailResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "admissionNumber": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "currentClass": {
            "$ref": "#/components/schemas/CurrentClass"
          },
          "enrollmentDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParentInfo"
            }
          },
          "feeSummary": {
            "$ref": "#/components/schemas/FeeSummary"
          },
          "medicalNotes": {
            "type": "string"
          },
          "profilePhotoUrl": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListStudentFeeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentFeeResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseListFeeStructureResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeeStructureResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FeeStructureResponse": {
        "type": "object",
        "properties": {
          "structureId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "termName": {
            "type": "string"
          },
          "sessionName": {
            "type": "string"
          },
          "totalAmount": {
            "type": "number"
          },
          "mandatoryAmount": {
            "type": "number"
          },
          "applicableToClasses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "applicableClassCount": {
            "type": "integer",
            "format": "int32"
          },
          "studentCount": {
            "type": "integer",
            "format": "int32"
          },
          "collectionRate": {
            "type": "number",
            "format": "double"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseFeeDashboardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FeeDashboardResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FeeDashboardResponse": {
        "type": "object",
        "properties": {
          "termName": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/DashboardSummary"
          },
          "byClass": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassCollection"
            }
          },
          "upcomingDeadlines": {
            "$ref": "#/components/schemas/UpcomingDeadlines"
          },
          "dailyCollectionTrend": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyTrend"
            }
          }
        }
      },
...
      "ApiResponsePageResponseUserSummaryResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseUserSummaryResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}